fix: enable healthcheck with missing identity
This patch fixes an issue that caused the lack of the health check
against the provider API server when access was configured through a
control plane secret. With this fixwhen the ForeignCluster controller
does not found an Identity resource for the given cluster it looks
for control plane secrets to set the APIServer URL in the status,
used then for the health check.